requirements.txt
altgraph==0.17.4
certifi==2025.7.9
charset-normalizer==3.4.2
dnspython==2.7.0
idna==3.10
packaging==25.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
python-dateutil==2.9.0.post0
python-whois==0.9.5
requests==2.32.4
six==1.17.0
urllib3==2.5.0
